backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
qdrant-client>=1.7.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pyjwt>=2.8.0
bcrypt>=4.0.0
email-validator>=2.0.0
